tests: fix CellularData_1 flaky test
The test switches the flight mode multiple times. If one switch the
flight mode on and don't wait for it to come up, it'll mess up the next
time it switches the flight mode.

Setting the correct expectation on the flight mode switch make sure it
waits for the state transition to complete. This avoids a race condition
the next time the mode is switched and fixes this test.