nmofono: Avoid WifiLinks of unmanaged adapters
On the Pixel 3a testing shows that, under some circumstances,
the list of APs cannot be filled due to WifiLinkItem::Private::updateAccessPoints
not being emitted. This is due to the indicator code hooking up
to signals of the unmanaged wlan1 adapter instead of the allowed wlan0 one.

To fight this issue ensure we only update WifiLinks of managed adapters.