nmofono: have SimManager update the stored SIM info as it changes
It can happen that the information stored in QSettings for a SIM is
missing the IMSI number; this seems to happen in Android 9 devices,
where the IMSI number is not delivered immediately as the SIM is
reported, but sometime later.

With this change we keep listening for changes of the SIM properties and
update the stored data accordingly. Therefore the stored data will
always have the IMSI number, and at startup time we will have all the
needed information to setup the SIMs.